+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added arthropods
 - A 'point light glow' effect, making lanterns and other point lights more visually pronounced
 - Generate random name for site2 sites
+- Shader dithering to remove banding from scenes with large colour gradients
 
 ### Changed